% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PKPDdatasets.R
\docType{data}
\name{dapa_IV_oral}
\alias{dapa_IV_oral}
\title{IV and oral pharmacokinetic data for daptomycin}
\format{A data frame with 1536 rows and 11 variables}
\usage{
data(dapa_IV_oral)
}
\description{
A dataset containing simulated dapagliflozin PK.  A single IV dose followed by 3 escalating oral doses
\itemize{
  \item ID. Numerical ID  (1-24)
  \item TIME. Nominal Time after first dose (hrs)
  \item TAD. Time After Dose (hrs)
  \item COBS. Observed Concentration (ug/L)
  \item AMT_IV. IV dose amount when given (ug)
  \item AMT_ORAL. Oral dose amount when given (ug)
  \item OCC. Occasion, associated with each dosing event
  \item AGE. Age (years)
  \item WEIGHT Weight (kg)
  \item GENDER. Gender flag (female=1, male=0)
  \item FORMULATION Formulation associated with dose (IV or oral)
}
}
\keyword{datasets}
