% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{pain_relief}
\alias{pain_relief}
\title{Simulated anti-inflammatory pain trial}
\format{A data frame with 1920 rows and 7 variables}
\usage{
data(pain_relief)
}
\description{
CTM_Nopain is a novel anti-inflammatory agent under preliminary investigation.
A dose-ranging trial was conducted comparing placebo with 3 doses of CTM_Nopain (5mg, 20mg and 80 mg QD).
The maximum tolerated dose is 160 mg per day.
Plasma concentrations (mg/L) of the drug were measured at 0, 0.5, 1, 1.5, 2, 2.5, 3-8 hours.
Pain score (0=no pain, 1=mild, 2=moderate, 3=severe) were obtained at time points when plasma concentration was collected.
A pain score of 2 or more is considered as no pain relief.
The model used is a one-compartment PK model with allometric model of weight on CL (exp 0.75) and volume (exp 1)
with direct PD effect (Emax model).
}
\details{
\itemize{
  \item ARM Factor with 4 levels
  \item ID. Numerical ID  (1-160)
  \item TIME Time in hours
  \item CONC Concentration (mg/L)
  \item PAINRELIEF 0 (no relief) or 1 (relief)
  \item PAINSCORE categorical pain score
  \item DOSE Dose Given
}
}
\keyword{datasets}

