% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PKPDdatasets.R
\docType{data}
\name{sd_oral_richpk}
\alias{sd_oral_richpk}
\title{One-compartment pharmacokinetic data given single oral dose}
\format{A data frame with 4300 rows and 9 variables}
\usage{
data(sd_oral_richpk)
}
\description{
A dataset containing dose, plasma concentration and time data, as well as
 demographic data of Age, Weight, Gender, and Race for each individual
}
\details{
\itemize{
  \item ID. Numerical ID  (1--50)
  \item Time. Time in hours (0--24)
  \item Amt. Amount of drug given, time dependent, in milligrams
  \item Conc. Plasma concentration in mg/L
  \item Age. Age in years
  \item Weight. Weight in kg
	 \item Gender. Male or Female gender identification
  \item Race. Ethnicity
  \item Dose. Dose given to each individual in milligrams
}
}
\keyword{datasets}

